requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
textblob==0.18.0.post0
nltk==3.9.1
tqdm==4.66.5
snscrape==0.7.0.20230622
